requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
APScheduler==3.10.4
pyserial==3.5
pytest==7.4.0
pytest-cov==4.1.0
playwright==1.40.0
pytest-playwright==0.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
pymodbus>=3.0.0
psutil>=5.8.0
pytest-xdist==3.5.0
requests==2.31.0